requirements.txt
fastapi==0.116.1
pydantic==2.11.7
uvicorn==0.35.0
python-dotenv==1.1.1
requests==2.32.4
langchain==0.3.27
langchain-google-genai==2.1.9
markitdown==0.1.2
chromadb==1.0.17
langchain_text_splitters==0.3.9
langchain_chroma==0.2.5
langgraph==0.6.5
ipython==9.4.0
gradio==5.44.1